requirements.txt
# Core Frameworks
Django==3.2.18
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-environ==0.10.0

# Asynchronous Tasks
celery==5.3.4
redis==4.5.5
django-celery-beat==2.5.0

# Authentication & Security
djangorestframework-simplejwt==5.2.2
django-axes==6.1.0
bcrypt==4.0.1

# Database Management
psycopg2-binary==2.9.6
django-filter==23.2
django-import-export==3.3.6
drf-yasg==1.21.6

# Real-Time Monitoring
channels<=4.0.0
channels-redis==3.3.1
django-prometheus


# Third-Party Integrations
requests==2.31.0
python-dotenv==1.0.0
# stripe==5.7.0

# Deployment & Monitoring
gunicorn==21.2.0
supervisor==4.2.4
sentry-sdk==1.30.0
newrelic==9.2.0

# Testing & Code Quality
pytest==7.4.2
pytest-django==4.5.2
factory_boy==3.3.0
coverage==7.3.1
black==23.7.0
flake8==6.1.0
pre-commit==3.4.0

# Localization & Language Support
django-modeltranslation==0.18.11
Babel==2.12.1
